[JT-77] refactor: validatePaymentInfo 위치 변경
