+++ module-application/app-api/src/main/java/shop/jtoon/payment/application/PaymentInfoService.java
@@ -5,11 +5,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.jtoon.entity.Member;
 import shop.jtoon.entity.PaymentInfo;
+import shop.jtoon.exception.DuplicatedException;
 import shop.jtoon.payment.request.PaymentReq;
 import shop.jtoon.payment.response.PaymentRes;
 import shop.jtoon.repository.PaymentInfoRepository;
 import shop.jtoon.repository.PaymentInfoSearchRepository;
 import shop.jtoon.service.PaymentInfoDomainService;
+import shop.jtoon.type.ErrorStatus;
 
 import java.util.List;
 
@@ -40,11 +42,21 @@ public class PaymentInfoService {
     }
 
     public void validatePaymentInfo(PaymentReq paymentReq) {
-        paymentInfoDomainService.validatePaymentInfo(
-                paymentReq.impUid(),
-                paymentReq.impUid(),
-                paymentReq.itemName(),
-                paymentReq.amount()
-        );
+        paymentInfoDomainService.validatePaymentInfo(paymentReq.itemName(), paymentReq.amount());
+        validateImpUid(paymentReq.impUid());
+        validateMerchantUid(paymentReq.merchantUid());
+    }
+
+
+    private void validateImpUid(String impUid) {
+        if (paymentInfoRepository.existsByImpUid(impUid)) {
+            throw new DuplicatedException(ErrorStatus.PAYMENT_IMP_UID_DUPLICATED);
+        }
+    }
+
+    private void validateMerchantUid(String merchantUid) {
+        if (paymentInfoRepository.existsByMerchantUid(merchantUid)) {
+            throw new DuplicatedException(ErrorStatus.PAYMENT_MERCHANT_UID_DUPLICATED);
+        }
     }
 }
